Introduce registry pattern for Display types
Registry pattern will allow for new Display types to be more easily added in the future.
To add a new Display type:
- Define it as a new React component in `src/components/displayTypes/`.
- Ensure it uses the common props defined in `DisplayTypeRegistry.tsx`.
- Add it to the registry in `DisplayTypeRegistry.tsx`, associated with its string label.

- Replaced switch statement in `DisplayRenderer.tsx` with discrete definitions for each display type.
These are in `components/displayTypes/`. Currently have definitions for `ControlText`, `ControlBoolean`, and `ControlRange`.

- New display type definitions are registered in `DisplayTypeRegistry.tsx`.
This is the file that used to be `DisplayRenderer.tsx`. It still holds the typedef for the
props used by all Display types. Types are registered with a Record<string, React.FC> object.

- Updated `DiagramElement.tsx` to fetch the type of Display Component from the new registry. If no valid type exists, uses a yellow placeholder error Display.