Convert App into a Component: use JSON + draw CDDs
To prep for adding svelte-jsoneditor to this to convert it into an updated version of the CDD Authoring Tool,
I converted App.tsx into CausalDecisionDiagram.tsx.

CausalDecisionDiagram.tsx:
- Takes a JSON object useState (the CDM it is rendering and simulating).
For now this is never set to a new value, but when hooked to JSON editor, that will change.
- Now mostly holds React structures and creates the high-level HTML structure of a CDD

src/lib/evaluateModel.tsx:
- Moved the big evaluate model function out of `CausalDecisionDiagram.tsx` to here.
- This will make it easier to unit test this later, and more clearly separates concerns.
- CausalDecisionDiagram specifies the way state is handled in a CDD and constructs JSX,
while evaluateModel.tsx holds the evaluation logic for simulating a model